requirements.txt
psycopg2-binary
openai
python-dotenv
requests
